working led matrix foundation

+++ include/LEDManager.h
@@ -14,7 +14,7 @@ protected:
     float_t brightnessChange;
 
 public:
-	CLEDManager(uint16_t size, uint8_t brightness);
+	CLEDManager(uint16_t size, float brightness);
 
 #ifdef LED
     virtual uint16_t LED_Status(CRGB *leds) { return 0; };
