+++ src/LEDManager.cpp
@@ -2,7 +2,7 @@
 #include "LEDManager.h"
 #include "Configuration.h"
 
-CLEDManager::CLEDManager(uint16_t size, uint8_t brigthness) {
+CLEDManager::CLEDManager(uint16_t size, float brigthness) {
     this->size = size;
     this->brightness = brigthness;
 }
